requirements.txt
edge-tts>=6.1.0
aiohttp_socks>=0.8.4
PySide6
psutil>=6.0.0
python-dotenv
google-genai>=1.0.1
